Fix: 이메일 양식 수정 + PB 회원 가입시 기본 profile등록 + PB삭제시 오류 해결 (#184)
* Fix: 이메일 알림시 양식 수정 + PB 삭제시 오류 해결

* Fix: test 코드 수정

---------

+++ src/main/java/kr/co/moneybridge/core/util/S3Util.java
@@ -75,6 +75,7 @@ public class S3Util {
     // s3에서 파일 삭제
     public void delete(String profile) {
         int index = cloudFrontDomain.length();
+        if(profile.length() < index) return;
         String fileName = profile.substring(index + 1);
 
         s3Client.deleteObject(new DeleteObjectRequest(bucket, fileName));
